intensity overflow when too big

+++ celltk/utils/track_utils.py
@@ -19,7 +19,7 @@ def calc_diff(a, b):
 
 
 def calc_massdiff(cells0, cells1):
-    return calc_ratiodiff([i.total_intensity for i in cells0], [i.total_intensity for i in cells1])
+    return calc_ratiodiff([i.total_intensity.astype(np.float32) for i in cells0], [i.total_intensity.astype(np.float32) for i in cells1])
 
 
 def find_one_to_one_assign(cost):
